refactor: reorganize Home page into modular and reusable components
