+++ frontend/index.html
@@ -1,12 +1,14 @@
 <!doctype html>
-<html lang="en">
+<html lang="pt-BR">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" href="/leaf-green-icon.svg" type="image/svg+xml" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>frontend</title>
+    <title>EcoPoints </title>
+    
+    <meta name="description" content="Encontre pontos de coleta, registre descartes e faça parte da mudança. Reciclagem fácil e impacto real." />
   </head>
-  <body>
+  <body class="bg-surface">
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
