+++ frontend/src/App.css
@@ -1,3 +1,5 @@
+/*remover dps????*/
+
 #root {
   max-width: 1280px;
   margin: 0 auto;
