+++ frontend/src/components/layout/Footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <footer className="bg-gray-900 text-gray-400 py-10 text-center">
+      <p>© 2025 EcoPoints. Todos os direitos reservados.</p>
+    </footer>
+  );
+}
